Restructure CCA to have 1:M relationships
