+++ src/main/java/org/ohdsi/webapi/cohortcomparison/CohortInfo.java
@@ -1,0 +1,18 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package org.ohdsi.webapi.cohortcomparison;
+
+import java.io.Serializable;
+
+/**
+ *
+ * @author asena5
+ */
+public class CohortInfo implements Serializable {
+	public int id;
+	public String name;
+	public String expression;
+}
